requirements.txt
langchain==0.2.3
langchain_community==0.2.4
langchain_core==0.2.5
langchain_huggingface==0.0.3
python-dotenv==1.0.1
streamlit==1.35.0
streamlit_chat==0.1.1